+++ components/scroller/modules/core-scroller.vue
@@ -46,6 +46,10 @@
   </div>
 </template>
 
+<!--
+This is a code from external library (https://github.com/Akryum/vue-virtual-scroller/blob/master/packages/vue-virtual-scroller/src/components/RecycleScroller.vue)
+We have modified it for our own specific use.
+-->
 <script setup>
 import { computed, watch, markRaw, shallowReactive, onMounted, nextTick, reactive, ref } from 'vue';
