+++ components/scroller/modules/dynamic-scroller.vue
@@ -33,6 +33,10 @@
   </core-scroller>
 </template>
 
+<!--
+This is a code from external library (https://github.com/Akryum/vue-virtual-scroller/blob/master/packages/vue-virtual-scroller/src/components/DynamicScroller.vue)
+We have modified it for our own specific use.
+-->
 <script>
 import CoreScroller from './core-scroller.vue';
 import DtScrollerItem from './scroller-item.vue';
